services/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Authentication Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "uvloop==0.19.0",
    "sqlalchemy==2.0.27",
    "pydantic[email]==2.6.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "psycopg2-binary==2.9.9",
    "python-decouple==3.8",
    "alembic==1.13.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["auth"]
